move our files into an extension directory
