+++ src/extensions/replay/sessionrecording-utils.ts
